ingest/pyproject.toml
[project]
name = "ingest"
version = "0.1.0"
description = "QTM file ingestion based on project settings"
requires-python = ">=3.8"
dependencies = [
    "mysql-connector-python",
    "pyyaml",
    "python-dotenv",
    "sshtunnel",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"